Phase 02 finished
The results of the K-D tree's getNearest and a linear search of the
array are identical when given a random point. The test was onducted
100,000 times and the two methods agreed 100% of the time. The
getNearest method seems to be asymptotically faster than the linear
search as it takes around 0.01 seconds for both to exicute 1000 times
but takes getNearest only around 2 seconds to run 100,000 times while
the linear search requires more than 30 seconds.